+++ ssbd02/src/main/java/pl/lodz/p/it/ssbd2021/ssbd02/ejb/mok/facades/interfaces/AccountFacadeLocal.java
@@ -48,7 +48,7 @@ public interface AccountFacadeLocal extends AbstractFacadeInterface<Account> {
      *                    w przypadku braku jego aktywacji.
      * @return Lista obiektów typu {@link Account}.
      */
-    List<Account> findByUnconfirmedAndExpired(int removalTime);
+    List<Account> findByUnconfirmedAndExpired(long removalTime);
 
     /**
      * Metoda wyszukująca listę encji typu {@link Account} o przekazanym adresie e-mail.
